please be the last time

+++ dash/dl_model.py
@@ -3,9 +3,9 @@ from pathlib import Path
 
 import logging
 
-logging.critical('Beginning file download with urllib2...')
 
-url = 'https://docs.google.com/uc?export=download&id=1x__YR6PO6_cW-vy3vb0cdjWRYqjBqPW0'
-urllib.request.urlretrieve(url, "./model.pkl")
 
-logging.critical('Finished download')
+def download():
+    if not (Path.cwd()/"model.pkl").exists()
+        url = 'https://docs.google.com/uc?export=download&id=1x__YR6PO6_cW-vy3vb0cdjWRYqjBqPW0'
+        urllib.request.urlretrieve(url, "./model.pkl")
